feat: gw: Send out "status" events with !accepted
that is on invalid token provided lol
previously it just closed the websocket connection

which is kinda bad cuz the websockets api doesnt
really provide a proper, non-hackish way to differentiate
between the websocket connection being closed due to it shitting
itself or due to it being closed by the remote

if there is however a way i will just find a reason to not remove
it. no, i am not using timing nor the lack of messages sent to
check for that. thank you very much.

Signed-off-by: Emilia Luminé <eqilia@national.shitposting.agency>